Revert "16 add gnosis logo and dex nfts dao pages (#18)"
This reverts commit dd0c40796358eb87d5ddd4c9ab052062b0d90037.
