+++ next.config.js
@@ -11,18 +11,9 @@ const nextConfig = {
         protocol: 'https',
         hostname: 'example.com',
         port: '',
+        // pathname: '/account123/**',
       },
     ],
-  },
-  output: 'export',
-  trailingSlash: true,
-  exportPathMap: function () {
-    return {
-      '/': { page: '/' },
-      '/dex': { page: '/dex' },
-      '/nfts': { page: '/nfts' },
-      '/dao': { page: '/dao' },
-    }
   },
 }
