+++ src/components/layout/Header.tsx
@@ -1,10 +1,9 @@
 import React from 'react'
-import { Flex, useColorModeValue, Spacer, Heading, Text } from '@chakra-ui/react'
+import { Flex, useColorModeValue, Spacer, Heading } from '@chakra-ui/react'
 import { SITE_NAME } from '../../utils/config'
 import { LinkComponent } from './LinkComponent'
 import { ThemeSwitcher } from './ThemeSwitcher'
 import { ConnectKitButton } from 'connectkit'
-import Image from 'next/image'
 
 interface Props {
   className?: string
@@ -24,24 +23,6 @@ export function Header(props: Props) {
       <Spacer />
 
       <Flex alignItems="center" gap={4}>
-        <LinkComponent href="./dex">
-          <Heading as="h2" size="s">
-            DEX
-          </Heading>
-        </LinkComponent>
-        <LinkComponent href="./nfts">
-          <Heading as="h2" size="s">
-            NFTs
-          </Heading>
-        </LinkComponent>
-        <LinkComponent href="./dao">
-          <Heading as="h2" size="s">
-            DAO
-          </Heading>
-        </LinkComponent>
-        <LinkComponent target="blank" href="https://www.gnosis.io/">
-          <Image priority height="30" width="30" alt="contract-image" src="./gnosis-owl.png" />
-        </LinkComponent>
         <ConnectKitButton />
         <ThemeSwitcher />
       </Flex>
